Compute a layout for subfeature data

Read subfeature data such as domains or blast hits on genes into a tidy dataframe. Subfeatures need to be associated with an already added feature track. The subfeature track itself is internally converted into a new, regular feature track by mapping the `start` and `end` coordinates provided relative to their parent feature into coordinates relative to the sequences underlying the parent features.

as_subfeatures(x, seqs, features, ..., everything = TRUE)

Arguments

x
subfeature data convertible to a feature layout

seqs
the sequence layout the parent features map onto.

features
the parent features the subfeatures map onto.

...
passed on to `layout_seqs()`

everything
set to FALSE to drop optional columns

transform
use if features and subfeatures are in different coordinate spaces, i.e. if matching nucleotide-level annotations to protein level annotations, e.g. genes and protein blast results.

Value

a tbl_df with plot coordinates

Details

Obligatory columns are `feature_id`, `start` and `end`. Also recognized are `strand` and `bin_id`.

Note `start` and `end` for every record will be coerced so that `start < end`. If no `strand` was provided, `strand` will be added and set to "+" for records that initially had `start < end` and "-" for `end < start` inputs. If `strand` was provided, `start` and `end` will be reorganized to conform with `start < end` without any additional effect.