Defined file formats and extensions

For seamless reading of different file formats, gggenomes uses a mapping of known formats to associated file extensions and contexts in which the different formats can be read. The notion of context allows one to read different information from the same format/extension. For example, a gbk file holds both feature and sequence information. If read in "feats" context read_feats("*.gbk") it will return a feature table, if read in "seqs" context read_seqs("*.gbk"), a sequence index.

file_formats(context)

Arguments

context
a file format context defined in gggenomes_global$file_formats

Value

dictionarish vector of file formats with recognized extensions as names

Defined Contexts, Formats and Extensions

feats    ambigious  [txt,tsv,csv]
         bed      [bed]
         blast    [m8,o6,o7]
         fasta    [fa,fas,fasta,ffn,fna,faa]
         gbk      [gbk,gb,gbff]
         gff3     [gff,gff3]
         paf      [paf]
seqs     ambigious  [txt,tsv,csv]
         fai      [fai]
         seq_len  [fa,fas,fasta,ffn,fna,faa,gff,gbk]
zips     bz2      [bz,bz2]
         gz       [gz]
         xz       [xz]
         zip      [zip]

Examples

# vector of defined zip formats and recognized extensions as names
file_formats("zips")

#>    bz   bz2    gz    xz   zip 
#> "bz2" "bz2"  "gz"  "xz" "zip"