Draw bin labels

Put bin labels left of the sequences. `nudge_left` adds space relative to the total bin width between the label and the seqs, by default 5 expands the plot to the left by 20

geom_bin_label(mapping = NULL, data = use_bins(), hjust = 1,
  size = 6, nudge_left = 0.05, expand_left = 0.2, expand_x = NULL,
  expand_aes = NULL, ...)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

nudge_left
by this much relative to the widest bin

expand_left
by this much relative to the widest bin

expand_x
expand the plot to include this absolute x value

expand_aes
provide custom aes mappings for the expansion (advanced)

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

Details

Set `x` and `expand_x` to an absolute position to align all labels at a specific location