Do numeric values fall into specified ranges?

Do numeric values fall into specified ranges?

in_range(x, left, right, closed = TRUE)

Arguments

x
a numeric vector of values

left, right
boundary values or vectors of same length as x

closed
wether to include (`TRUE`) or exclude (`FALSE`) the endpoints. Provide 2 values for different behaviors for lower and upper boundary, e.g. `c(TRUE, FALSE)` to include only the lower boundary.

Value

a logical vector of the same length as the input

Examples