Read features and links from common file formats

Read features or links from common formats, such as GFF3, Genbank, BED, BLAST tabular output or PAF files. File formats and the format-specific read_*() function are automatically determined based in file extensions, if possible. Can read multiple files in the same format into a single table: useful, for example, to read a folder of gff-files with each containing genes of a different genome.

read_feats(files, format = NULL, .id = "file_id", ...)

read_subfeats(files, format = NULL, .id = "file_id", ...)

read_links(files, format = NULL, .id = "file_id", ...)

read_sublinks(files, format = NULL, .id = "file_id", ...)

Arguments

files
files to reads. Should all be of same format.

format
If NULL, will be guessed from file extension. Else, any format known to gggenomes ("gff3", "gbk", ... see file_formats() for full list) or any suffix that maps to a known read_<suffix> function to be called, such as readr::read_tsv(), for example.

.id
When binding output from several files, how to name the column with the name of the file each record came from. Defaults to "file_id". Set to "bin_id" if every file represents a different bin.

...
additional arguments passed on to the specific read function for the given format.

Value

a gggenomes compatible feature or link tibble

Examples

# read a file
read_feats(ex("eden-utr.gff"))

#> Reading as gff3:

#> * eden-utr [/home/runner/work/_temp/Library/gggenomes/extdata/eden-utr.gff]

#> Harmonizing column names
#> * ID -> feat_id
#> * Parent -> parent_ids
#> * Name -> name

#> Features read

#> # A tibble: 8 x 3
#>   source type                n
#>   <chr>  <chr>           <int>
#> 1 NA     cDNA_match          1
#> 2 NA     CDS                 5
#> 3 NA     exon                5
#> 4 NA     five_prime_UTR      1
#> 5 NA     gene                1
#> 6 NA     mRNA                5
#> 7 NA     TF_binding_site     1
#> 8 NA     three_prime_UTR     1

#> # A tibble: 20 x 15
#>    file_id seq_id start   end strand type  feat_id introns parent_ids source
#>    <chr>   <chr>  <int> <int> <chr>  <chr> <chr>   <list>  <list>     <chr> 
#>  1 eden-u… ctg123  1050  9000 +      cDNA… match0… <int [… <chr [1]>  NA    
#>  2 eden-u… ctg123  1201  7600 +      CDS   cds000… <int [… <chr [1]>  NA    
#>  3 eden-u… ctg123  1201  7600 +      CDS   cds000… <int [… <chr [1]>  NA    
#>  4 eden-u… ctg123  3301  7600 +      CDS   cds000… <int [… <chr [1]>  NA    
#>  5 eden-u… ctg123  3391  7600 +      CDS   cds000… <int [… <chr [1]>  NA    
#>  6 eden-u… ctg123  1201  7600 +      CDS   cds000… <int [… <chr [1]>  NA    
#>  7 eden-u… ctg123  1300  1500 +      exon  exon00… <NULL>  <chr [1]>  NA    
#>  8 eden-u… ctg123  1050  1500 +      exon  exon00… <NULL>  <chr [2]>  NA    
#>  9 eden-u… ctg123  3000  3902 +      exon  exon00… <NULL>  <chr [2]>  NA    
#> 10 eden-u… ctg123  5000  5500 +      exon  exon00… <NULL>  <chr [3]>  NA    
#> 11 eden-u… ctg123  7000  9000 +      exon  exon00… <NULL>  <chr [3]>  NA    
#> 12 eden-u… ctg123  1050  1200 +      five… NA      <NULL>  <chr [1]>  NA    
#> 13 eden-u… ctg123  1000  9000 +      gene  gene00… <NULL>  <chr [1]>  NA    
#> 14 eden-u… ctg123  1050  9000 +      mRNA  mRNA00… <int [… <chr [1]>  NA    
#> 15 eden-u… ctg123  1050  9000 +      mRNA  mRNA00… <int [… <chr [1]>  NA    
#> 16 eden-u… ctg123  1300  9000 +      mRNA  mRNA00… <int [… <chr [1]>  NA    
#> 17 eden-u… ctg123  1300  9000 +      mRNA  mRNA00… <int [… <chr [1]>  NA    
#> 18 eden-u… ctg123  1050  9000 +      mRNA  mRNA00… <int [… <chr [1]>  NA    
#> 19 eden-u… ctg123  1000  1012 +      TF_b… tfbs00… <NULL>  <chr [1]>  NA    
#> 20 eden-u… ctg123  7601  9000 +      thre… NA      <NULL>  <chr [1]>  NA    
#> # … with 5 more variables: score <chr>, phase <int>, name <chr>, Target <chr>,
#> #   geom_id <chr>

if (FALSE) {
# read all gffs from a directory
read_feats(list.files("path/to/directory", "*.gff$", full.names=TRUE))
# read remote files
gbk_phages <- c(
PSSP7 = "ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/858/745/GCF_000858745.1_ViralProj15134/GCF_000858745.1_ViralProj15134_genomic.gff.gz",
PSSP3 = "ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/904/555/GCF_000904555.1_ViralProj195517/GCF_000904555.1_ViralProj195517_genomic.gff.gz")
read_feats(gbk_phages)
}