Do numeric values fall into specified ranges?

Do numeric values fall into specified ranges?

in_range(x, left, right, closed = TRUE)

Arguments

x
a numeric vector of values

left, right
boundary values or vectors of same length as x

closed
wether to include (`TRUE`) or exclude (`FALSE`) the endpoints. Provide 2 values for different behaviors for lower and upper boundary, e.g. `c(TRUE, FALSE)` to include only the lower boundary.

Value

a logical vector of the same length as the input

Examples

in_range(1:5, 2, 4)

#> Error in in_range(1:5, 2, 4): could not find function "in_range"

in_range(1:5, 2, 4, closed=c(FALSE, TRUE)) # left-open

#> Error in in_range(1:5, 2, 4, closed = c(FALSE, TRUE)): could not find function "in_range"

in_range(1:5, 6:2, 3) # vector of boundaries, single values recycle

#> Error in in_range(1:5, 6:2, 3): could not find function "in_range"

# plays nicely with dplyr
df <- tibble(x=rep(4,5), left=1:5, right=3:7)
mutate(df,
  closed=in_range(x, left, right, TRUE),
  open=in_range(x, left, right, FALSE))

#> Error in in_range(x, left, right, TRUE): could not find function "in_range"