Compute a genomes layout from sequences, features and links

See `gggenomes::gggenomes()` for more info.

layout_genomes(seqs = NULL, genes = NULL, features = NULL,
  links = NULL, infer_bin_id = seq_id, infer_start = min(start, end),
  infer_end = max(start, end), infer_length = max(start, end), ...)

Arguments

seqs
a table with sequence data (seq_id, bin_id, length)

genes
a table or a list of table with gene data to be added as feature tracks. Required columns: seq_id, bin_id, start, end.

For a single table, adds the track_id will be "genes". For a list, track_ids are parsed from the list names, or if names are missing from the name of the variable containing each table.

features
same as genes, but the single table track_id will default to "features".

links
a table or a list of tables with link data to be added as link tracks (columns: from, to, from_start, from_end, to_start, to_end). Same naming scheme as for features.

infer_bin_id, infer_start, infer_end, infer_length
used to infer pseudo seqs if only features or links are provided. The expressions are evaluated in the context of the first feature track, or the first links track.

By default subregions of sequences from the first to the last feature/link are generated. Set `infer_start` to 0 to show all sequences from their true beginning.

...
layout parameters passed on to `layout_seqs()`

features_track_id
default ID for first features track

links_track_id
default ID for first links track