Read features from GFF3 files

Files with ##FASTA section work but result in parsing problems for all lines of the fasta section. Just ignore those warnings, or strip the fasta section ahead of time from the file.

read_gff3(
  file,
  sources = NULL,
  types = NULL,
  infer_cds_parents = FALSE,
  col_names = def_names("gff3"),
  col_types = def_types("gff3")
)

Arguments

sources
only return features from these sources

types
only return features of these types, e.g. gene, CDS, ...

infer_cds_parents
infer the mRNA parent for CDS features based on overlapping coordinates. In most GFFs this is properly set, but sometimes this information is missing. Generally, this is not a problem, however, geom_gene calls parse the parent information to determine which CDS and mRNAs are part of the same gene model. Without the parent info, mRNA and CDS are plotted as individual features.

col_names
column names to use. Defaults to def_names("blast") compatible with blast tabular output (--outfmt 6/7 in blast++ and -m8 in blast-legacy). def_names("blast") can easily be combined with extra columns: col_names = c(def_names("blast"), "more", "things").

Value

tibble